@@ -12,41 +12,40 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "pack": "cd ./dist && npm pack"
   },
-  "dependencies": {
-    "@angular/common": "5.2.0",
-    "@angular/compiler": "5.2.0",
-    "@angular/compiler-cli": "5.2.0",
-    "@angular/core": "5.2.0",
-    "@angular/forms": "5.2.0",
-    "@angular/http": "5.2.0",
-    "@angular/platform-browser": "5.2.0",
-    "@angular/platform-browser-dynamic": "5.2.0",
-    "@angular/router": "5.2.0",
-    "core-js": "2.4.1",
-    "rxjs": "5.5.6",
-    "zone.js": "0.8.19"
+  "peerDependencies": {
+    "@angular/common": "6.0.0",
+    "@angular/core": "6.0.0"
   },
   "devDependencies": {
-    "@types/jasmine": "2.8.3",
-    "@types/jasminewd2": "2.0.2",
-    "@types/node": "6.0.60",
-    "codelyzer": "4.0.1",
-    "del": "3.0.0",
-    "gulp": "github:gulpjs/gulp#4.0",
-    "jasmine-core": "2.8.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular-devkit/architect": "0.6.0",
+    "@angular-devkit/build-ng-packagr": "0.6.0",
+    "@angular/cli": "6.0.0",
+    "ng-packagr": "3.0.0-rc.2",
+    "@types/jasmine": "2.8.6",
+    "@types/jasminewd2": "2.0.3",
+    "@types/node": "8.9.4",
+    "codelyzer": "4.2.1",
+    "jasmine-core": "2.99.1",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
+    "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.2.1",
-    "karma-jasmine": "1.1.0",
+    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-jasmine": "1.1.1",
     "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.1.2",
-    "rename": "1.0.4",
-    "ts-node": "4.1.0",
+    "ts-node": "5.0.1",
+    "tslib": "1.7.1",
     "tslint": "5.9.1",
-    "typescript": "2.5.3"
+    "typescript": "2.7.2",
+    "core-js": "2.5.4",
+    "rxjs": "6.0.0",
+    "zone.js": "0.8.26"
   }
 }
