@@ -1,6 +1,3 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { ApiModule } from './api.module';
-
-platformBrowserDynamic().bootstrapModule(ApiModule)
-  .catch(err => console.log(err));
+export * from './api.module';
+export * from './api.service';
+export * from './dtos';
